[feat] : 객체스토리지 관련 수정
