+++ src/main/resources/application-prod.yml
@@ -15,3 +15,9 @@ spring:
       hibernate:
         format_sql: true
         show_sql: true
+
+  # 객체 스토리지(GCS, S3 등)
+  config:
+    storage:
+      bucket-name: bucket-name
+      credential-path: classpath:service-account-key.json
